backend/requirements.txt
Flask==3.0.0
flask-cors
Pillow